Clean up loading Indicator demo
